add new views (#1731)
* add new views

* wip test

* margin tests

* more tests'

* remove only

* rename func

* naming fix

+++ markets/perps-market/contracts/interfaces/IPerpsMarketModule.sol
@@ -32,7 +32,7 @@ interface IPerpsMarketModule {
 
     function indexPrice(uint128 marketId) external view returns (uint);
 
-    function fillPrice(uint128 marketId, int orderSize, uint price) external returns (uint);
+    function fillPrice(uint128 marketId, int128 orderSize, uint price) external returns (uint);
 
     /**
      * @dev Given a marketId return a market's summary details in one call.
